+++ shop/Controllers/SuppliersController.cs
@@ -26,7 +26,7 @@ namespace shop.Controllers
         }
 
         // GET: Suppliers/Details/5
-        public async Task<IActionResult> Details(long? id)
+        public async Task<IActionResult> Details(int? id)
         {
             if (id == null || _context.Suppliers == null)
             {
@@ -88,7 +88,7 @@ namespace shop.Controllers
         }
 
 
-        public async Task<IActionResult> Edit(long? id)
+        public async Task<IActionResult> Edit(int? id)
         {
             if (id == null || _context.Suppliers == null)
             {
@@ -114,7 +114,7 @@ namespace shop.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(long id, [Bind("Id,Name,Phone,Email,Address")] Supplier supplier,bool State)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,Name,Phone,Email,Address")] Supplier supplier,bool State)
         {
             if (id != supplier.Id)
             {
@@ -155,7 +155,7 @@ namespace shop.Controllers
         }
 
         // GET: Suppliers/Delete/5
-        public async Task<IActionResult> Delete(long? id)
+        public async Task<IActionResult> Delete(int? id)
         {
             var sup = await _context.Suppliers.FindAsync(id);
             if (sup != null)
@@ -183,7 +183,7 @@ namespace shop.Controllers
         // POST: Suppliers/Delete/5
         //[HttpPost, ActionName("Delete")]
         //[ValidateAntiForgeryToken]
-        //public async Task<IActionResult> DeleteConfirmed(long id)
+        //public async Task<IActionResult> DeleteConfirmed(int id)
         //{
         //    if (_context.Suppliers == null)
         //    {
@@ -199,7 +199,7 @@ namespace shop.Controllers
         //    return RedirectToAction(nameof(Index));
         //}
 
-        private bool SupplierExists(long id)
+        private bool SupplierExists(int id)
         {
           return _context.Suppliers.Any(e => e.Id == id);
         }
